user-story-tests/03-web-component-from-another-entity.html

<!DOCTYPE html>
<html lang="en">

<head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
    <title>Web component from another entity</title>
    <script src='./resources/testharness.js'></script>
    <script src='./resources/testharnessreport.js'></script>
</head>

<body>

    <script type="module">
        import { WebComponentFromAnotherEntity } from "./components/WebComponentFromAnotherEntity.js";
        customElements.define(
            "web-component-from-another-entity",
            WebComponentFromAnotherEntity
        );
    </script>
    <style>
        @layer resets {
            button {
                border: thick dashed red;
            }
        }
    </style>
    <web-component-from-another-entity>
        <template shadowrootmode="open">
            <style>
                @layer inherit.resets;
            </style>
        </template>
    </web-component-from-another-entity>

    <script type="module">
        import ShadowLayers from "../ShadowLayers.js";
        import { adoptStylesFromLayerStatementRule } from "../AdoptStyles.js";

        window.onload = function () {

            const elName = "web-component-from-another-entity"

            test(function () {

                const inheritResult = adoptStylesFromLayerStatementRule(document.querySelector(elName))
                assert_true(inheritResult, "inherit successfully returns");

                const layerStatementResult = ShadowLayers.findLayerStatementRule(document.querySelector(elName))
                assert_equals(typeof layerStatementResult, "object", "found, so returns object");

                const layerStatementRule = layerStatementResult.styleSheet.cssRules.item(layerStatementResult.ruleIndex)
                assert_array_equals(["inherit.resets"],
                    layerStatementRule.nameList,
                    "@layer resets");

            }, "Web component from another entity");

        }
    </script>

    <div id="log"></div>

</body>

</html>
